Unit Details : Allow aspect ratios to be specified for images, to stop flickers
When changing modifications on a unit ( or a unit itself ), the scroll jumps
and images flicker. This is because they do not have explicit heights.
So, when an image is swapped out for another, the image container collapses
until the image is fetched and loaded and then comes back up again.